% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiemBaseExt-package.R
\docType{package}
\name{RiemBaseExt-package}
\alias{RiemBaseExt-package}
\title{Extension to 'RiemBase' Package}
\description{
We provide a number of off-the-shelf algorithms for clustering, elementary computation, and others. 
See Bhattacharya and Bhattacharya (2012) <doi:10.1017/CBO9781139094764> for more details regarding Statistics on Manifolds.
}
